# docs/openapi/replicadb-server.json
{
  "components": {
    "responses": {
      "BadRequestProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "The request is malformed or fails validation."
      },
      "ConflictProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "The request conflicts with the current resource or run state."
      },
      "ForbiddenProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "The authenticated identity lacks the required permission."
      },
...
      "NotFoundProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "The requested resource does not exist or is not visible."
      },
...
      "UnauthorizedProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "Authentication is required or credentials are invalid."
      }
    },
    "schemas": {
...
      "DatasourceCapabilitiesResponse": {
        "description": "Role-specific connector capabilities used to validate job bindings.",
        "properties": {
          "singleJobOnly": {
            "description": "Whether the connector requires jobs=1.",
            "type": "boolean"
          },
          "sinkCapable": {
            "description": "Whether the connector can receive sink rows.",
            "type": "boolean"
          },
          "sinkModes": {
            "items": {
              "description": "Supported sink mode.",
              "enum": [
                "complete",
                "complete-atomic",
                "incremental"
              ],
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "sourceCapable": {
            "description": "Whether the connector can provide source rows.",
            "type": "boolean"
          },
          "sourceModes": {
            "items": {
              "description": "Supported source mode.",
              "enum": [
                "complete",
                "complete-atomic",
                "incremental"
              ],
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "sourceQuery": {
            "description": "Whether source.query is supported.",
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "DatasourcePermissionRequest": {
        "description": "Complete replacement set of grants for one user and datasource.",
        "properties": {
          "permissions": {
            "items": {
              "description": "Datasource permission.",
              "enum": [
                "VIEW",
                "USE",
                "EDIT",
                "VIEW",
                "USE",
                "EDIT"
              ],
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          }
        },
        "required": [
          "permissions"
        ],
        "type": "object"
      },
      "DatasourcePermissionResponse": {
        "description": "Datasource grants grouped for one user.",
        "properties": {
          "permissions": {
            "items": {
              "description": "Granted datasource permission.",
              "enum": [
                "VIEW",
                "USE",
                "EDIT",
                "VIEW",
                "USE",
                "EDIT"
              ],
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "userId": {
            "description": "Granted user identifier.",
            "format": "uuid",
            "type": "string"
          },
          "username": {
            "description": "Granted user's current username.",
            "type": "string"
          }
        },
        "type": "object"
      },
      "DatasourceRequest": {
        "description": "Request for a managed datasource profile. Unknown fields are rejected.",
        "properties": {
          "clearSecurityKeys": {
            "description": "Security keys to remove explicitly during an update.",
            "example": [
              "password"
            ],
            "items": {
              "description": "Security keys to remove explicitly during an update.",
              "example": "[\"password\"]",
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "connectorType": {
            "description": "Registered connector wire name.",
            "example": "postgres",
            "type": "string"
          },
          "name": {
            "description": "Unique display name.",
            "example": "Warehouse source",
            "type": "string"
          },
          "security": {
            "additionalProperties": {
              "description": "Security values to encrypt. On update, omitted or blank entries preserve stored values.",
              "type": "string",
              "writeOnly": true
            },
            "description": "Security values to encrypt. On update, omitted or blank entries preserve stored values.",
            "type": "object",
            "writeOnly": true
          },
          "technicalParams": {
            "additionalProperties": {
              "description": "Non-secret connector settings. Omitted values default to an empty map.",
              "type": "string"
            },
            "description": "Non-secret connector settings. Omitted values default to an empty map.",
            "type": "object"
          }
        },
        "required": [
          "connectorType",
          "name"
        ],
        "type": "object"
      },
      "DatasourceResponse": {
        "description": "Redacted managed datasource profile and permission-aware capabilities.",
        "properties": {
          "canEdit": {
            "description": "Whether the current identity may edit this profile.",
            "type": "boolean"
          },
          "canUse": {
            "description": "Whether the current identity may bind this profile to a job.",
            "type": "boolean"
          },
          "canView": {
            "description": "Whether the current identity may view this profile.",
            "type": "boolean"
          },
          "capabilities": {
            "$ref": "#/components/schemas/DatasourceCapabilitiesResponse"
          },
          "connectorType": {
            "description": "Registered connector wire name.",
            "type": "string"
          },
          "createdAt": {
            "description": "Creation timestamp in UTC.",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Datasource identifier.",
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "description": "Datasource display name.",
            "type": "string"
          },
          "safeConnectDisplay": {
            "description": "Credential-redacted connection display suitable for the UI.",
            "type": "string"
          },
          "securityConfigured": {
            "description": "Whether encrypted security material is configured, without exposing it.",
            "type": "boolean"
          },
          "technicalParams": {
            "additionalProperties": {
              "description": "Non-secret connector settings.",
              "type": "string"
            },
            "description": "Non-secret connector settings.",
            "type": "object"
          },
          "updatedAt": {
            "description": "Last update timestamp in UTC.",
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "JobPermissionRequest": {
        "description": "Complete replacement set of grants for one user and job.",
        "properties": {
          "permissions": {
            "items": {
              "description": "Job permission.",
              "enum": [
                "VIEW",
                "EDIT",
                "EXECUTE",
                "CANCEL",
                "VIEW",
                "EDIT",
                "EXECUTE",
                "CANCEL"
              ],
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          }
        },
        "required": [
          "permissions"
        ],
        "type": "object"
      },
      "JobPermissionResponse": {
        "description": "Job grants grouped for one user.",
        "properties": {
          "permissions": {
            "items": {
              "description": "Granted job permission.",
              "enum": [
                "VIEW",
                "EDIT",
                "EXECUTE",
                "CANCEL",
                "VIEW",
                "EDIT",
                "EXECUTE",
                "CANCEL"
              ],
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "userId": {
            "description": "Granted user identifier.",
            "format": "uuid",
            "type": "string"
          },
          "username": {
            "description": "Granted user's current username.",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "RFC 7807 problem response. Dynamic detail text is credential-redacted.",
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "format": "uri-reference",
            "type": "string"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "format": "uri-reference",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "csrfHeader": {
        "description": "CSRF header for protected state-changing requests. Its value matches the XSRF-TOKEN cookie.",
        "in": "header",
        "name": "X-XSRF-TOKEN",
        "type": "apiKey"
      },
      "sessionCookie": {
        "description": "Session cookie issued by an authenticated, same-origin ReplicaDB Server deployment.",
        "in": "cookie",
        "name": "JSESSIONID",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Static contract for the authenticated ReplicaDB Server control plane. Use it against a same-origin deployment that owns the session and CSRF cookies.",
    "title": "ReplicaDB Server API",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/datasources/{datasourceId}/permissions": {
      "get": {
        "description": "Returns grants grouped by user. ADMIN is required.",
        "operationId": "listDatasourcePermissions",
        "parameters": [
          {
            "description": "Datasource identifier.",
            "in": "path",
            "name": "datasourceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasourcePermissionResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Datasource grants returned"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundProblem"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List datasource permissions",
        "tags": [
          "Datasource permissions"
        ]
      }
    },
    "/api/v1/datasources/{datasourceId}/permissions/{userId}": {
      "delete": {
        "description": "Removes all datasource grants for one user. ADMIN and CSRF are required.",
        "operationId": "revokeDatasourcePermissions",
        "parameters": [
          {
            "description": "Datasource identifier.",
            "in": "path",
            "name": "datasourceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Target user identifier.",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Datasource grants revoked"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundProblem"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "Revoke datasource permissions",
        "tags": [
          "Datasource permissions"
        ]
      },
      "put": {
        "description": "Atomically replaces one user's VIEW, USE, and EDIT grant set. ADMIN and CSRF are required.",
        "operationId": "replaceDatasourcePermissions",
        "parameters": [
          {
            "description": "Datasource identifier.",
            "in": "path",
            "name": "datasourceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Target user identifier.",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasourcePermissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DatasourcePermissionResponse"
                }
              }
            },
            "description": "Datasource grants replaced"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestProblem"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundProblem"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "Replace datasource permissions",
        "tags": [
          "Datasource permissions"
        ]
      }
    },
    "/api/v1/datasources/{id}": {
      "delete": {
        "description": "Deletes an unreferenced datasource. ADMIN and CSRF are required; profiles bound to jobs return a conflict.",
        "operationId": "deleteDatasource",
        "parameters": [
          {
            "description": "Datasource identifier.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Datasource deleted"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundProblem"
          },
          "409": {
            "$ref": "#/components/responses/ConflictProblem"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "Delete a datasource",
        "tags": [
          "Datasources"
        ]
      },
      "get": {
        "description": "Returns redacted profile metadata and capability flags after VIEW permission. Stored security values are never rehydrated in the response.",
        "operationId": "getDatasource",
        "parameters": [
          {
            "description": "Datasource identifier.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DatasourceResponse"
                }
              }
            },
            "description": "Datasource returned"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundProblem"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get a datasource",
        "tags": [
          "Datasources"
        ]
      },
      "put": {
        "description": "Updates a profile after EDIT permission. Omitted or blank security inputs preserve stored encrypted values; clearSecurityKeys explicitly removes named values. CSRF is required.",
        "operationId": "updateDatasource",
        "parameters": [
          {
            "description": "Datasource identifier.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DatasourceResponse"
                }
              }
            },
            "description": "Datasource updated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestProblem"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundProblem"
          },
          "409": {
            "$ref": "#/components/responses/ConflictProblem"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "Update a datasource",
        "tags": [
          "Datasources"
        ]
      }
    },
...
    "/api/v1/jobs/{jobDefinitionId}/permissions/{userId}": {
      "delete": {
        "description": "Removes all job grants for one user after EDIT permission. CSRF is required.",
        "operationId": "revokeJobPermissions",
        "parameters": [
          {
            "description": "Job definition identifier.",
            "in": "path",
            "name": "jobDefinitionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Target user identifier.",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Job grants revoked"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundProblem"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "Revoke job permissions",
        "tags": [
          "Job permissions"
        ]
      },
      "put": {
        "description": "Atomically replaces one user's VIEW, EDIT, EXECUTE, and CANCEL grants after EDIT permission. CSRF is required.",
        "operationId": "replaceJobPermissions",
        "parameters": [
          {
            "description": "Job definition identifier.",
            "in": "path",
            "name": "jobDefinitionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Target user identifier.",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobPermissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/JobPermissionResponse"
                }
              }
            },
            "description": "Job grants replaced"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestProblem"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundProblem"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "Replace job permissions",
        "tags": [
          "Job permissions"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ],
...
}